+++ dom/examples/todo/src/item.rs
@@ -9,8 +9,9 @@ use moxie_dom::{
 };
 
 #[topo::nested]
-#[illicit::from_env(todos: Key<Vec<Todo>>)]
+#[illicit::from_env(todos: &Key<Vec<Todo>>)]
 fn item_edit_input(todo: Todo, editing: Key<bool>) -> Input {
+    let todos = todos.clone();
     let text = todo.text.clone();
     text_input(&text, true, move |value: String| {
         editing.set(false);
@@ -25,8 +26,9 @@ fn item_edit_input(todo: Todo, editing: Key<bool>) -> Input {
 }
 
 #[topo::nested]
-#[illicit::from_env(todos: Key<Vec<Todo>>)]
+#[illicit::from_env(todos: &Key<Vec<Todo>>)]
 fn item_with_buttons(todo: Todo, editing: Key<bool>) -> Div {
+    let todos = todos.clone();
     let id = todo.id;
     let toggle_todos = todos.clone();